Don't explicitly enable CFI in libhevc.
Previously there was no other sanitization, so in order for the
blacklist to be picked up CFI had to be enabled. With integer_overflow
enabled, we no longer need it enabled explicitly in the build file.
Since libhevc is a static library, both CFI and non-CFI variants will
be built and linked accordingly.

If libhevc is used as a shared library in the future, we'll need to
explicitly re-enable CFI in the build file.

Bug: 63926619
Test: Library builds.
Change-Id: If852aa732d12f1176f57c3c9f7f4574d709c13ab