libhevcenc: Move deblk, sao and subpel plane generation into nctb loop
Currently deblocking, sao and subpel plane generation are happening for every
CTB serially. This change enables deblocking to be finished for the entire row first.
Then performs SAO for the entire row and finally performs subpel plane
generation for the entire row. This change does not have any effect on the
end bitstream but is expected to give performance gains on devices with
low instruction cache.

On arm platforms, it is noticed that for P5 preset, performance is improved by 15%.

Test: hevcenc -c vid_enc_cfg.txt
Bug: 135510351

Change-Id: I6d09e6cc1d36e967a9b3dd3e68f75bdd0b476a1c